Refactor bot architecture and update dependencies
The architecture of the bot was significantly refactored, including a shift from a singleton PrismaClient to a class-based approach with the PrismaClient as a class property. Moreover, several project dependencies got updated such as Prisma and typescript. The outdated logging library (winston) was replaced with a new one (pino), and essential bot settings were moved to a separate config file. The changes aim to enhance maintainability and readability in the code.